python-worker/requirements.txt
# python-worker/requirements.txt
pymongo==4.6.0
torch==2.1.2
torchaudio==2.1.2
pyannote.audio==3.1.1
faster-whisper==1.0.0
soundfile==0.12.1
librosa==0.10.1
numpy==1.24.3
python-dotenv==1.0.0

